pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "porter"
version = "0.1.0"
description = "Dual-pane terminal file manager for homelabs and sysadmins"
requires-python = ">=3.11"
dependencies = [
    "textual>=8.0",
    "paramiko>=3.0",
    "PyYAML>=6.0",
]

[project.scripts]
porter = "porter.__main__:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["porter*"]
